requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ca-certificates=2020.12.5=h5b45459_0
certifi=2020.12.5=py39hcbf5309_1
click=7.1.2=pyh9f0ad1d_0
cmake=3.18.4.post1=pypi_0
dlib=19.22.0=py39hf8509d4_0
face_recognition=1.3.0=pyhd3deb0d_2
face_recognition_models=0.3.0=pyh9f0ad1d_0
fastapi=0.63.0=pyhd8ed1ab_0
freetype=2.10.4=h546665d_1
h11=0.12.0=pyhd8ed1ab_0
intel-openmp=2021.2.0=h57928b3_616
jpeg=9d=h8ffe710_0
lcms2=2.12=h2a16943_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
liblapack=3.9.0=8_mkl
libpng=1.6.37=h1d00b33_2
libtiff=4.2.0=hc10be44_1
lz4-c=1.9.3=h8ffe710_0
mkl=2020.4=hb70f87d_311
numpy=1.20.2=py39h6635163_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.5.1.48=pypi_0
openjpeg=2.4.0=h48faf41_0
openssl=1.1.1k=h8ffe710_0
pillow=8.1.2=py39h1a9d4f7_1
pip=21.1.1=pyhd8ed1ab_0
pydantic=1.8.1=py39hb82d6ee_1
python=3.9.2=h7840368_0_cpython
python-dotenv=0.17.1=pyhd8ed1ab_0
python-multipart=0.0.5=py_0
python_abi=3.9=1_cp39
pyyaml=5.4.1=py39hb82d6ee_0
setuptools=49.6.0=py39hcbf5309_3
six=1.15.0=pyh9f0ad1d_0
sqlite=3.35.5=h8ffe710_0
starlette-base=0.13.6=py_0
tk=8.6.10=h8ffe710_1
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
tzdata=2021a=he74cb21_0
uvicorn=0.13.4=py39hcbf5309_0
vc=14.2=hb210afc_4
vs2015_runtime=14.28.29325=h5e1d092_4
watchgod=0.6=py_0
websockets=8.1=py39hb82d6ee_3
wheel=0.36.2=pyhd3deb0d_0
wincertstore=0.2=py39hcbf5309_1006
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
zlib=1.2.11=h62dcd97_1010
zstd=1.4.9=h6255e5f_0
